refactor: fix some pylint issues
