+++ python/introduction/if_else.py
@@ -8,9 +8,9 @@ def main(n: int) -> None:
         print("Weird")
     # n is even
     else:
-        if n >= 2 and n <= 5:
+        if 2 <= n <= 5:
             print("Not Weird")
-        elif n >= 6 and n <= 20:
+        elif 6 <= n <= 20:
             print("Weird")
         else:
             print("Not Weird")
