+++ python/regex_and_parsing/html_parser2.py
@@ -1,7 +1,10 @@
 """ HTML Parser - Part 2 """
+# pylint: disable=duplicate-code
+
 from abc import ABC
 from html.parser import HTMLParser
-from typing import List
+
+from .utils import Reader
 
 # create a subclass and override the handler methods
 
@@ -17,24 +20,6 @@ class MyHTMLParser(HTMLParser, ABC):
     def handle_data(self, data: str) -> None:
         if data != '\n':
             print(f">>> Data\n{data}")
-
-
-class Reader():
-    """ Reader """
-    MAX_LINES = 100
-
-    def readlines(self) -> List[str]:
-        """ Read lines """
-        number_of_lines = int(input())
-        if number_of_lines > self.MAX_LINES:
-            raise ValueError("Number of lines is too big")
-        lines = []
-        for _ in range(number_of_lines):
-            lines.append(input()+"\n")
-        return lines
-
-
-# instantiate the parser and fed it some HTML
 
 
 def main() -> None:
