+++ python/sets/difference.py
@@ -1,5 +1,6 @@
 # Enter your code here. Read input from STDIN. Print output to STDOUT
 """ Set .difference() Operation """
+# pylint: disable=duplicate-code
 
 if __name__ == "__main__":
